fixed win issues

+++ include/vigra/multi_blockwise.hxx
@@ -72,7 +72,7 @@ namespace blockwise{
             //std::cout<<"blockshape "<<(*iter).core().size()<<"\n";
 
             #pragma omp for
-            for(size_t i=0 ; i<blocking.numBlocks(); ++i){
+            for(int i=0 ; i<blocking.numBlocks(); ++i){
 
                 const BlockWithBorder bwb = iter[i];
